Refactor/only 4

+++ arduino/digital_i_o.go
@@ -15,16 +15,16 @@ const (
 //
 //go:wasm-module arduino
 //export digitalRead
-func DigitalRead(pin uint8) int32
+func DigitalRead(pin int32) int32
 
 // Equivalent to Arduino's digitalWrite https://www.arduino.cc/reference/en/language/functions/digital-io/digitalwrite/
 //
 //go:wasm-module arduino
 //export digitalWrite
-func DigitalWrite(pin uint8, value int32)
+func DigitalWrite(pin int32, value int32)
 
 // Equivalent to Arduino's pinMode https://www.arduino.cc/reference/en/language/functions/digital-io/pinmode/
 //
 //go:wasm-module arduino
 //export pinMode
-func PinMode(pin uint8, mode uint8)
+func PinMode(pin int32, mode int32)
